refactor : Component 별 패키지 분리
